src/testing/test_brand.py
```python
import unittest
from src.brand.brand import Brand

class TestBrand(unittest.TestCase):

    def setUp(self):
        self.brand = Brand()

    def test_create_brand(self):
        brand_name = "Test Brand"
        self.brand.create_brand(brand_name)
        self.assertEqual(self.brand.name, brand_name)

    def test_create_branded_ai_agent(self):
        brand_name = "Test Brand"
        ai_agent_name = "Test AI Agent"
        self.brand.create_brand(brand_name)
        self.brand.create_branded_ai_agent(ai_agent_name)
        self.assertEqual(self.brand.ai_agents[0].name, ai_agent_name)

    def test_license_branded_ai_agent(self):
        brand_name = "Test Brand"
        ai_agent_name = "Test AI Agent"
        self.brand.create_brand(brand_name)
        self.brand.create_branded_ai_agent(ai_agent_name)
        license = self.brand.license_branded_ai_agent(ai_agent_name)
        self.assertIsNotNone(license)

if __name__ == '__main__':
    unittest.main()
```